# lib/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /withdrawals:
    get:
      operationId: getWithdrawals
      tags: [Withdrawals]
      summary: Get user's withdrawal history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Withdrawal list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalListResponse'
    post:
      operationId: createWithdrawal
      tags: [Withdrawals]
      summary: Submit a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWithdrawalRequest'
      responses:
        '201':
          description: Withdrawal submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Withdrawal'
  /withdrawals/{id}:
    delete:
      operationId: deleteWithdrawal
      tags: [Withdrawals]
      summary: Delete a withdrawal (processed or rejected only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /withdrawals/{id}/process:
    put:
      operationId: processWithdrawal
      tags: [Withdrawals]
      summary: Admin process a withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Withdrawal processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Withdrawal'

  # ─── Coupons ─────────────────────────────────────────────────
...
components:
  schemas:
...
    Withdrawal:
      type: object
      required: [id, userId, amount, phone, country, operator, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        amount:
          type: number
        phone:
          type: string
        country:
          type: string
        operator:
          type: string
        code:
          type: string
          nullable: true
          description: Withdrawal code obtained by the user from 1xBet
        status:
          type: string
          enum: [pending, processed, rejected]
        createdAt:
          type: string
          format: date-time
    CreateWithdrawalRequest:
      type: object
      required: [amount, phone, country, operator, code]
      properties:
        amount:
          type: number
        phone:
          type: string
        country:
          type: string
        operator:
          type: string
        code:
          type: string
          description: Withdrawal code obtained by the user from 1xBet
    WithdrawalListResponse:
      type: object
      required: [withdrawals, total, page, limit]
      properties:
        withdrawals:
          type: array
          items:
            $ref: '#/components/schemas/Withdrawal'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string
...